experiments/requirements.txt
torch
torchvision
diffusers
transformers
accelerate
hydra-core
omegaconf
wandb
datasets
torchmetrics[image]
openai
tqdm
Pillow
